requirements.txt
luigi==3.5.0
pandas==2.0.3
pangres==4.2.1
python-dotenv==1.0.1
Requests==2.31.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9